Changed name of the project and executable from iREVEAL to iRevealLit
