+++ Java/make_ireveallite_jar.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+javac -cp gson-2.2.4.jar DataModel/*java
+jar xf gson-2.2.4.jar
+jar cmf MANIFEST.MF iRevealLite.jar com/* DataModel/*.class
