--- examples/iReveal.csv
@@ -1,22 +1,0 @@
-I/O data for Boiler_ROM,Input,Input,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output,Output
-Case,TotalMassFlow_SecGP,TotalMassFlow_OFAGP,Temperature_FlueGP,O2_MassFlow_FlueGP,N2_MassFlow_FlueGP,H2O_MassFlow_FlueGP,CO2_MassFlow_FlueGP,CO_MassFlow_FlueGP,SO2_MassFlow_FlueGP,HCl_MassFlow_FlueGP,Temperature_FlySP0,C_FlySP0,H_FlySP0,N_FlySP0,Cl_FlySP0,S_FlySP0,O_FlySP0,Ash_FlySP0,moisture_FlySP0,heat_loss,Qplaten
-1,2.9604E+02,1.2134E+02,1.4997E+03,1.9033E+01,3.9300E+02,2.9499E+01,1.1911E+02,2.2365E-03,6.3312E-01,5.1726E-03,1.4997E+03,6.7110E-02,4.6204E-03,1.2279E-03,1.0400E-05,6.5559E-04,9.6258E-03,5.6146E+00,0.0000E+00,5.6579E+08,1.0079E+08
-2,3.1372E+02,1.1295E+02,1.4957E+03,2.1110E+01,4.0006E+02,2.9602E+01,1.1918E+02,2.0184E-03,6.3350E-01,5.1757E-03,1.4957E+03,4.7953E-02,3.3014E-03,8.7741E-04,7.4400E-06,4.6845E-04,6.8780E-03,5.6146E+00,0.0000E+00,5.5938E+08,9.9115E+07
-3,2.8780E+02,1.1200E+02,1.5080E+03,1.5076E+01,3.7964E+02,2.9308E+01,1.1900E+02,2.7969E-03,6.3257E-01,5.1681E-03,1.5080E+03,9.5622E-02,6.5834E-03,1.7496E-03,1.4800E-05,9.3412E-04,1.3715E-02,5.6146E+00,0.0000E+00,5.7786E+08,1.0428E+08
-4,3.1854E+02,1.2789E+02,1.4864E+03,2.5615E+01,4.1507E+02,2.9804E+01,1.1924E+02,1.6201E-03,6.3375E-01,5.1777E-03,1.4864E+03,3.4843E-02,2.3988E-03,6.3753E-04,5.4000E-06,3.4038E-04,4.9976E-03,5.6146E+00,0.0000E+00,5.4563E+08,9.5541E+07
-5,3.1243E+02,1.2550E+02,1.4902E+03,2.3682E+01,4.0862E+02,2.9717E+01,1.1921E+02,1.7714E-03,6.3362E-01,5.1767E-03,1.4902E+03,4.1569E-02,2.8620E-03,7.6062E-04,6.4500E-06,4.0609E-04,5.9624E-03,5.6146E+00,0.0000E+00,5.5161E+08,9.7004E+07
-6,2.7291E+02,1.1979E+02,1.5128E+03,1.3504E+01,3.7426E+02,2.9227E+01,1.1892E+02,3.1481E-03,6.3219E-01,5.1650E-03,1.5128E+03,1.1493E-01,7.9129E-03,2.1030E-03,1.7800E-05,1.1228E-03,1.6485E-02,5.6146E+00,0.0000E+00,5.8149E+08,1.0616E+08
-7,2.8530E+02,1.2224E+02,1.5049E+03,1.6824E+01,3.8552E+02,2.9391E+01,1.1904E+02,2.5470E-03,6.3277E-01,5.1697E-03,1.5049E+03,8.5503E-02,5.8867E-03,1.5645E-03,1.3300E-05,8.3527E-04,1.2264E-02,5.6146E+00,0.0000E+00,5.7201E+08,1.0284E+08
-8,2.7950E+02,1.1742E+02,1.5102E+03,1.4441E+01,3.7746E+02,2.9274E+01,1.1896E+02,2.9406E-03,6.3239E-01,5.1666E-03,1.5102E+03,1.0462E-01,7.2027E-03,1.9142E-03,1.6200E-05,1.0220E-03,1.5006E-02,5.6146E+00,0.0000E+00,5.7913E+08,1.0509E+08
-9,3.0447E+02,1.2758E+02,1.4928E+03,2.2348E+01,4.0414E+02,2.9655E+01,1.1919E+02,1.8868E-03,6.3349E-01,5.1756E-03,1.4928E+03,4.8225E-02,3.3202E-03,8.8239E-04,7.4800E-06,4.7111E-04,6.9171E-03,5.6146E+00,0.0000E+00,5.5573E+08,9.8040E+07
-10,2.9905E+02,1.3032E+02,1.4941E+03,2.1749E+01,4.0211E+02,2.9625E+01,1.1917E+02,1.9439E-03,6.3339E-01,5.1748E-03,1.4941E+03,5.3286E-02,3.6686E-03,9.7500E-04,8.2600E-06,5.2055E-04,7.6430E-03,5.6146E+00,0.0000E+00,5.5749E+08,9.8533E+07
-11,2.7073E+02,1.1521E+02,1.5162E+03,1.1990E+01,3.6912E+02,2.9152E+01,1.1887E+02,3.4849E-03,6.3193E-01,5.1628E-03,1.5162E+03,1.2839E-01,8.8391E-03,2.3491E-03,1.9900E-05,1.2542E-03,1.8415E-02,5.6146E+00,0.0000E+00,5.8613E+08,1.0768E+08
-12,2.9478E+02,1.1616E+02,1.5027E+03,1.7577E+01,3.8811E+02,2.9430E+01,1.1907E+02,2.4188E-03,6.3296E-01,5.1712E-03,1.5027E+03,7.5618E-02,5.2061E-03,1.3836E-03,1.1700E-05,7.3870E-04,1.0846E-02,5.6146E+00,0.0000E+00,5.7030E+08,1.0200E+08
-13,3.0911E+02,1.3109E+02,1.4890E+03,2.4204E+01,4.1034E+02,2.9739E+01,1.1922E+02,1.7220E-03,6.3362E-01,5.1766E-03,1.4890E+03,4.1803E-02,2.8781E-03,7.6489E-04,6.4800E-06,4.0838E-04,5.9960E-03,5.6146E+00,0.0000E+00,5.5012E+08,9.6553E+07
-14,2.9020E+02,1.2357E+02,1.5017E+03,1.8227E+01,3.9026E+02,2.9459E+01,1.1908E+02,2.3467E-03,6.3298E-01,5.1714E-03,1.5017E+03,7.4773E-02,5.1480E-03,1.3682E-03,1.1600E-05,7.3046E-04,1.0725E-02,5.6146E+00,0.0000E+00,5.6794E+08,1.0153E+08
-15,2.8203E+02,1.0992E+02,1.5120E+03,1.3321E+01,3.7368E+02,2.9221E+01,1.1893E+02,3.1291E-03,6.3224E-01,5.1654E-03,1.5120E+03,1.1209E-01,7.7173E-03,2.0510E-03,1.7400E-05,1.0950E-03,1.6078E-02,5.6146E+00,0.0000E+00,5.8306E+08,1.0598E+08
-16,2.6685E+02,1.1822E+02,1.5175E+03,1.1794E+01,3.6845E+02,2.9142E+01,1.1886E+02,3.5766E-03,6.3189E-01,5.1625E-03,1.5175E+03,1.3023E-01,8.9660E-03,2.3828E-03,2.0200E-05,1.2722E-03,1.8679E-02,5.6146E+00,0.0000E+00,5.8613E+08,1.0813E+08
-17,2.7541E+02,1.3198E+02,1.5061E+03,1.6806E+01,3.8541E+02,2.9386E+01,1.1902E+02,2.5927E-03,6.3267E-01,5.1689E-03,1.5061E+03,9.0344E-02,6.2200E-03,1.6531E-03,1.4000E-05,8.8257E-04,1.2958E-02,5.6146E+00,0.0000E+00,5.7111E+08,1.0318E+08
-18,3.0391E+02,1.3289E+02,1.4905E+03,2.3433E+01,4.0775E+02,2.9703E+01,1.1920E+02,1.7865E-03,6.3354E-01,5.1760E-03,1.4905E+03,4.5656E-02,3.1433E-03,8.3539E-04,7.0800E-06,4.4601E-04,6.5485E-03,5.6146E+00,0.0000E+00,5.5246E+08,9.7155E+07
-19,2.6217E+02,1.2478E+02,1.5180E+03,1.2214E+01,3.6989E+02,2.9163E+01,1.1888E+02,3.5448E-03,6.3199E-01,5.1633E-03,1.5180E+03,1.2557E-01,8.6454E-03,2.2977E-03,1.9500E-05,1.2267E-03,1.8011E-02,5.6146E+00,0.0000E+00,5.8387E+08,1.0809E+08
-20,2.6621E+02,1.1350E+02,1.5199E+03,1.0597E+01,3.6438E+02,2.9082E+01,1.1881E+02,3.8819E-03,6.3167E-01,5.1607E-03,1.5199E+03,1.4168E-01,9.7542E-03,2.5923E-03,2.2000E-05,1.3840E-03,2.0321E-02,5.6146E+00,0.0000E+00,5.9002E+08,1.0925E+08
