--- iReveal/CCSI.h
@@ -1,19 +1,0 @@
-//CCSI.h
-
-#ifndef __CCSI_H__
-#define __CCSI_H__
-#ifndef NULL
-#define NULL 0
-#endif
-#define T_REAL double
-#define TINY 1.0e-20	//note: float range is [3.4e-38, 3.4e+38]
-#define PI 3.1415926535
-#define URGAS 8314.0
-#include <iostream>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <vector>
-using namespace std;
-
-#endif
