--- iReveal/Util.h
@@ -1,12 +1,0 @@
-//Util.h
-#include "CCSI.h"
-
-#ifndef __UTIL_H__
-#define __UTIL_H__
-
-string GetExecutableDirectory();
-void PrintCommandUsage();
-int ProcessJsonAndSampleInputSpace(char* filename);
-int BuildKrigingRom();
-
-#endif
