+++ iRevealLite.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "iRevealLite", "iRevealLite.vcxproj", "{BB231152-0090-4081-A159-2A774254B027}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Release|Win32 = Release|Win32
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BB231152-0090-4081-A159-2A774254B027}.Debug|Win32.ActiveCfg = Debug|Win32
+		{BB231152-0090-4081-A159-2A774254B027}.Debug|Win32.Build.0 = Debug|Win32
+		{BB231152-0090-4081-A159-2A774254B027}.Release|Win32.ActiveCfg = Release|Win32
+		{BB231152-0090-4081-A159-2A774254B027}.Release|Win32.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
