+++ iRevealLite/Util.h
@@ -1,0 +1,12 @@
+//Util.h
+#include "CCSI.h"
+
+#ifndef __UTIL_H__
+#define __UTIL_H__
+
+string GetExecutableDirectory();
+void PrintCommandUsage();
+int ProcessJsonAndSampleInputSpace(char* filename);
+int BuildKrigingRom();
+
+#endif
